backend/pyproject.toml
[tool.poetry]
name = "uurnik_connect_api"
version = "1.0"
description = ""
authors = ["Uurnik Systems <talha.javaid@uurnik.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nornir = "2.5"
ttp = "^0.6.0"
djangorestframework = "3.11.0"
django-cors-headers = "^3.6.0"
dnspython = "^2.0.0"
pysnmp = "^4.4.12"
beautifulsoup4 = "^4.9.3"
requests-ntlm2 = "^6.2.8"
mysqlclient = "^2.0.3"
requests = "^2.25.1"
nornir-scrapli = "2020.7.12"
scrapli = {extras = ["textfsm"], version = "^2020.12.31"}
scrapli-ssh2 = "^2020.10.24"
gunicorn = "^20.0.4"
djangorestframework-simplejwt = "^4.6.0"


[tool.poetry.dev-dependencies]
sphinx = "^3.2.1"
sphinx-rtd-theme = "^0.5.0"
recommonmark = "^0.6.0"
sphinx_markdown_tables = "^0.0.15"
rst2pdf = "^0.97"
black = "^21.5b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
